long instead of int

+++ src/main/java/com/pranavacharya/cses/WeirdAlgorithm.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 
 public class WeirdAlgorithm {
 
-    public void algo(int n) {
+    public void algo(long n) {
         StringBuilder sb = new StringBuilder();
         while (n != 1) {
             sb.append(n + " ");
@@ -21,7 +21,7 @@ public class WeirdAlgorithm {
     public static void main(String args[]) {
         WeirdAlgorithm wa = new WeirdAlgorithm();
         Scanner sc = new Scanner(System.in);
-        int n = sc.nextInt();
+        long n = sc.nextLong();
         wa.algo(n);
     }
